requirements.txt
# Core web framework for local API server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# ML / AI libraries
scikit-learn>=1.3.0
numpy>=1.24.0
scipy>=1.11.0

# Transformers / CodeBERT (optional heavy dependency)
transformers>=4.35.0
torch>=2.0.0

# NLP utilities
python-Levenshtein>=0.21.0

# CLI & terminal
click>=8.1.0
rich>=13.6.0

# Data persistence
# sqlite3 is stdlib — no install needed

# HTTP client (for VSCode extension ↔ server)
httpx>=0.25.0

# Code analysis
astroid>=3.0.0

# Data models / validation
pydantic>=2.4.0

# Packaging / distribution helpers
# (no extra deps needed beyond setuptools)
